move OpenAPIVerifier to a new module and update implementation
